refactor: split files, /src; feat: jest, first spec
