+++ .eslintrc.yml
@@ -2,3 +2,6 @@ extends: standard
 plugins:
   - standard
   - promise
+  - jest
+env:
+  jest/globals: true
