+++ package.json
@@ -2,9 +2,13 @@
   "name": "aegiseek",
   "version": "0.0.1",
   "license": "MIT",
-  "main": "server.js",
+  "main": "src/main.js",
   "scripts": {
-    "start": "node server.js"
+    "fix": "yarn lint --fix",
+    "lint": "eslint src",
+    "start": "node src/main.js",
+    "test": "jest --watch",
+    "test-once": "jest"
   },
   "dependencies": {
     "dotenv": "^6.0.0",
@@ -15,8 +19,10 @@
     "eslint": "^5.6.1",
     "eslint-config-standard": "^12.0.0",
     "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jest": "^21.23.0",
     "eslint-plugin-node": "^7.0.1",
     "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-standard": "^4.0.0"
+    "eslint-plugin-standard": "^4.0.0",
+    "jest": "^23.6.0"
   }
 }
